pyproject.toml
[tool.poetry]
name = "online_courses"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "3.2"
django-rest-framework = "^0.1.0"
poetry-dotenv-plugin = "^0.1.0-alpha.2"
django-debug-toolbar = "^3.2.4"
django-crispy-forms = "^1.14.0"
Pillow = "^9.0.1"
django-rq = "^2.5.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
